+++ Engine/Models/Location.cs
@@ -1,0 +1,104 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Xml.Linq;
+
+namespace Engine.Models
+{
+    public class Location
+    {
+        private int _xCoordinate;
+        private int _yCoordinate;
+        private string _name;
+        private string _description;
+        private string _imageName;
+
+        public Location(int xCoordinate, int yCoordinate, string name, string description, string imageName)
+        {
+            if (string.IsNullOrWhiteSpace(imageName) || string.IsNullOrWhiteSpace(description) || string.IsNullOrWhiteSpace(name))
+            {
+                throw new ArgumentException();
+            }
+
+            _name = name;
+            _xCoordinate = xCoordinate;
+            _yCoordinate = yCoordinate;
+            _description = description;
+            _imageName = "/Engine;component/Images/Locations/" + imageName;
+        }
+
+        public int XCoordinate
+        {
+            get
+            {
+                return _xCoordinate;
+            }
+            set
+            {
+                _xCoordinate = value;
+            }
+        }
+
+        public int YCoordinate
+        {
+            get
+            {
+                return _yCoordinate;
+            }
+            set
+            {
+                _yCoordinate = value;
+            }
+        }
+
+        public string Name
+        {
+            get
+            {
+                return _name;
+            }
+            set
+            {
+                if (string.IsNullOrWhiteSpace(value))
+                {
+                    throw new ArgumentException();
+                }
+                _name = value;
+            }
+        }
+
+        public string Description
+        {
+            get
+            {
+                return _description;
+            }
+            set
+            {
+                if (string.IsNullOrWhiteSpace(value))
+                {
+                    throw new ArgumentException();
+                }
+                _description = value;
+            }
+        }
+
+        public string ImageName
+        {
+            get
+            {
+                return _imageName;
+            }
+            set
+            {
+                if (string.IsNullOrWhiteSpace(value))
+                {
+                    throw new ArgumentException();
+                }
+                _imageName = value;
+            }
+        }
+    }
+}